# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Manuskript authoring and Story Oracle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video/projects/{projectId}/audio:
    post:
      operationId: queueAudioPass
      tags: [content-creators]
      summary: Queue an audio pass (noise reduction, EQ, ducking) for the SOUND leg
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoAudioPassInput" }
      responses:
        "201":
          description: Queued audio job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoJob" }
        "400":
          description: Invalid audio request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed (not the SOUND role / Captain)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/exports:
    post:
      operationId: queueVideoExport
      tags: [content-creators]
      summary: Queue multi-format exports (16:9, 9:16, 1:1) of the FINISH master
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoExportInput" }
      responses:
        "201":
          description: Queued export job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoJob" }
        "400":
          description: Invalid export request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed (not the FINISH role / Captain)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/thumbnail:
    post:
      operationId: queueVideoThumbnail
      tags: [content-creators]
      summary: Queue thumbnail frame extraction from the FINISH master
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoThumbnailInput" }
      responses:
        "201":
          description: Queued thumbnail job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoJob" }
        "400":
          description: Invalid thumbnail request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed (not the FINISH role / Captain)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/files/{fileId}/download:
    get:
      operationId: downloadVideoFile
      tags: [content-creators]
      summary: Download a file once the Lock is released (audited)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
        - { $ref: "#/components/parameters/VideoFileId" }
      responses:
        "200":
          description: File download stream
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "403":
          description: Lock is still on, or not a member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: File not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/downloads:
    get:
      operationId: listVideoDownloads
      tags: [content-creators]
      summary: List the download audit trail (Captain)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      responses:
        "200":
          description: Download audit trail
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VideoDownload" }
        "403":
          description: Captain only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  parameters:
...
    VideoProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    VideoFileId:
      name: fileId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VideoJob:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        assetId: { type: string }
        type: { type: string }
        status: { type: string }
        attempts: { type: integer }
        error: { type: ["string", "null"] }
        params: { type: ["object", "null"], additionalProperties: true }
        result: { type: ["object", "null"], additionalProperties: true }
        createdAt: { type: string, format: date-time }
        startedAt: { type: ["string", "null"], format: date-time }
        finishedAt: { type: ["string", "null"], format: date-time }
      required: [id, projectId, assetId, type, status, attempts, error, params, result, createdAt, startedAt, finishedAt]
...
    VideoAudioPassInput:
      type: object
      properties:
        action: { type: string, enum: [NOISE_REDUCTION, EQ, DUCKING, LEVELING] }
        assetId: { type: string }
      required: [action]
    VideoExportInput:
      type: object
      properties:
        formats:
          type: array
          items: { type: string, enum: ["16:9", "9:16", "1:1"] }
          minItems: 1
      required: [formats]
    VideoThumbnailInput:
      type: object
      properties:
        assetId: { type: string, minLength: 1 }
        timeMs: { type: integer, minimum: 0 }
      required: [assetId, timeMs]
    VideoDownload:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        fileId: { type: string }
        fileName: { type: string }
        memberId: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, fileId, fileName, memberId, createdAt]
...